+++ extras/AFC_prep.py
@@ -129,9 +129,9 @@ class afcPrep:
             extruder_obj.set_status_led( self.afc.led_tool_unloaded )
             if extruder_obj.on_shuttle():
                 # Calls ACTIVATE_EXTRUDER if current toolhead on shuttle is not the active extruder
-                if self.afc.function.get_current_extruder() != extruder_obj.name:
+                if self.afc.function.get_current_extruder() != extruder_obj.th_extruder_name:
                     self.afc.gcode.run_script_from_command(
-                        f"ACTIVATE_EXTRUDER EXTRUDER={extruder_obj.name}"
+                        f"ACTIVATE_EXTRUDER EXTRUDER={extruder_obj.th_extruder_name}"
                     )
             if 'system' in units and "extruders" in units["system"]:
                 # Check to see if lane_loaded is in dictionary and its not an empty string